# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenLearning API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/assessment/start": {
      "post": {
        "summary": "Assessment Start",
        "operationId": "assessment_start_api_assessment_start_post",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assessment/{session_id}/respond": {
      "post": {
        "summary": "Assessment Respond",
        "operationId": "assessment_respond_api_assessment__session_id__respond_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentRespondRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assessment/{session_id}/graph": {
      "get": {
        "summary": "Assessment Graph",
        "operationId": "assessment_graph_api_assessment__session_id__graph_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeGraphOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assessment/{session_id}/report": {
      "get": {
        "summary": "Assessment Report",
        "operationId": "assessment_report_api_assessment__session_id__report_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assessment/{session_id}/export": {
      "get": {
        "summary": "Assessment Export",
        "operationId": "assessment_export_api_assessment__session_id__export_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assessment/{session_id}/resume": {
      "get": {
        "summary": "Assessment Resume",
        "description": "Resume an active assessment session by loading the pending interrupt.",
        "operationId": "assessment_resume_api_assessment__session_id__resume_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssessmentReportResponse": {
        "properties": {
          "knowledgeGraph": {
            "$ref": "#/components/schemas/KnowledgeGraphOut"
          },
          "gapAnalysis": {
            "$ref": "#/components/schemas/EnrichedGapAnalysisOut"
          },
          "learningPlan": {
            "$ref": "#/components/schemas/LearningPlanOut"
          },
          "proficiencyScores": {
            "items": {
              "$ref": "#/components/schemas/ProficiencyScoreOut"
            },
            "type": "array",
            "title": "Proficiencyscores"
          }
        },
        "type": "object",
        "required": [
          "knowledgeGraph",
          "gapAnalysis",
          "learningPlan",
          "proficiencyScores"
        ],
        "title": "AssessmentReportResponse"
      },
      "AssessmentRespondRequest": {
        "properties": {
          "response": {
            "type": "string",
            "title": "Response"
          }
        },
        "type": "object",
        "required": [
          "response"
        ],
        "title": "AssessmentRespondRequest"
      },
      "AssessmentStartRequest": {
        "properties": {
          "skillIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skillids"
          },
          "targetLevel": {
            "type": "string",
            "title": "Targetlevel",
            "default": "mid"
          },
          "roleId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roleid"
          }
        },
        "type": "object",
        "required": [
          "skillIds"
        ],
        "title": "AssessmentStartRequest"
      },
      "AssessmentStartResponse": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "questionType": {
            "type": "string",
            "title": "Questiontype",
            "default": "calibration"
          },
          "step": {
            "type": "integer",
            "title": "Step",
            "default": 1
          },
          "totalSteps": {
            "type": "integer",
            "title": "Totalsteps",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "question"
        ],
        "title": "AssessmentStartResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgeGraphOut": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeNodeOut"
            },
            "type": "array",
            "title": "Nodes"
          }
        },
        "type": "object",
        "required": [
          "nodes"
        ],
        "title": "KnowledgeGraphOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}